requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.9
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.0
cron-descriptor==1.4.5
dill==0.4.0
Django==4.2
django-bootstrap-v5==1.0.11
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-multiselectfield==0.1.13
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
idna==3.10
isort==6.0.1
Jinja2==3.1.6
kombu==5.5.2
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psycopg==3.2.6
PyJWT==2.9.0
pylint==3.3.6
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
